pyproject.toml
[project]
name = "local-talking-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chatterbox-tts>=0.1.1",
    "langchain-ollama>=0.3.3",
    "langchain>=0.3.25",
    "nltk>=3.9.1",
    "openai-whisper>=20240930",
    "peft>=0.15.2",
    "rich>=14.0.0",
    "setuptools>=80.9.0",
    "sounddevice>=0.5.2",
    "torchaudio>=2.6.0",
    "nltk",
    "langchain-google-genai",
    "python-dotenv",
    "rich",
    "langchain-openai",
    "pyttsx3",
    "Flask",
]

[tool.setuptools.packages.find]
where = ["./"]
include = ["architect"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
